+++ app/src/main/java/de/triology/cas/oauth/web/CesOAuth20AccessTokenController.java
@@ -1,10 +1,28 @@
-package de.triology.cas.services.oauth;
+/*
+ * Licensed to Jasig under one or more contributor license
+ * agreements. See the NOTICE file distributed with this work
+ * for additional information regarding copyright ownership.
+ * Jasig licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file
+ * except in compliance with the License.  You may obtain a
+ * copy of the License at the following location:
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package de.triology.cas.oauth.web;
 
+import de.triology.cas.oauth.CesOAuthConstants;
+import de.triology.cas.oauth.CesOAuthUtils;
+import de.triology.cas.oauth.service.CesOAuthRegisteredService;
 import org.apache.commons.lang.StringUtils;
 import org.jasig.cas.services.ServicesManager;
-import org.jasig.cas.support.oauth.OAuthConstants;
-import org.jasig.cas.support.oauth.OAuthUtils;
-import org.jasig.cas.support.oauth.services.OAuthRegisteredService;
 import org.jasig.cas.ticket.ServiceTicket;
 import org.jasig.cas.ticket.TicketGrantingTicket;
 import org.jasig.cas.ticket.registry.TicketRegistry;
@@ -21,9 +39,9 @@ import javax.servlet.http.HttpServletResponse;
  * This controller returns an access token which is the CAS
  * granting ticket according to the service and code (service ticket) given.
  */
-public final class CesOAuthAccessTokenController extends AbstractController {
+public final class CesOAuth20AccessTokenController extends AbstractController {
 
-    private static Logger LOGGER = LoggerFactory.getLogger(CesOAuthAccessTokenController.class);
+    private static Logger LOGGER = LoggerFactory.getLogger(CesOAuth20AccessTokenController.class);
 
     private final ServicesManager servicesManager;
 
@@ -31,13 +49,8 @@ public final class CesOAuthAccessTokenController extends AbstractController {
 
     private final long timeout;
 
-    private static final String CONTENT_TYPE_JSON = "application/json";
-    private static final String TOKEN_TYPE = "token_type";
-    private static final String TOKEN_EXPIRES = "expires_in";
-    private static final String TOKEN_TYPE_VALUE = "Bearer";
-
-    public CesOAuthAccessTokenController(final ServicesManager servicesManager, final TicketRegistry ticketRegistry,
-                                         final long timeout) {
+    public CesOAuth20AccessTokenController(final ServicesManager servicesManager, final TicketRegistry ticketRegistry,
+                                           final long timeout) {
         this.servicesManager = servicesManager;
         this.ticketRegistry = ticketRegistry;
         this.timeout = timeout;
@@ -53,33 +66,33 @@ public final class CesOAuthAccessTokenController extends AbstractController {
     @Override
     protected ModelAndView handleRequestInternal(final HttpServletRequest request, final HttpServletResponse response) {
 
-        final String redirectUri = request.getParameter(OAuthConstants.REDIRECT_URI);
-        LOGGER.debug("{} : {}", OAuthConstants.REDIRECT_URI, redirectUri);
+        final String redirectUri = request.getParameter(CesOAuthConstants.REDIRECT_URI);
+        LOGGER.debug("{} : {}", CesOAuthConstants.REDIRECT_URI, redirectUri);
 
-        final String clientId = request.getParameter(OAuthConstants.CLIENT_ID);
-        LOGGER.debug("{} : {}", OAuthConstants.CLIENT_ID, clientId);
+        final String clientId = request.getParameter(CesOAuthConstants.CLIENT_ID);
+        LOGGER.debug("{} : {}", CesOAuthConstants.CLIENT_ID, clientId);
 
-        final String clientSecret = request.getParameter(OAuthConstants.CLIENT_SECRET);
+        final String clientSecret = request.getParameter(CesOAuthConstants.CLIENT_SECRET);
 
-        final String code = request.getParameter(OAuthConstants.CODE);
-        LOGGER.debug("{} : {}", OAuthConstants.CODE, code);
+        final String code = request.getParameter(CesOAuthConstants.CODE);
+        LOGGER.debug("{} : {}", CesOAuthConstants.CODE, code);
 
         //verify all components of the request are valid
         if (!verifyAccessTokenRequest(redirectUri, clientId, clientSecret, code)) {
-            return OAuthUtils.writeTextError(response, OAuthConstants.INVALID_REQUEST, 400);
+            return CesOAuthUtils.writeTextError(response, CesOAuthConstants.INVALID_REQUEST, 400);
         }
 
         final ServiceTicket serviceTicket = getServiceTicketFromTicketRegistry(code);
         if (serviceTicket == null) {
-            OAuthUtils.writeTextError(response, OAuthConstants.INVALID_GRANT, 400);
+            CesOAuthUtils.writeTextError(response, CesOAuthConstants.INVALID_GRANT, 400);
         }
 
         JSONObject json = makeJsonResponse(serviceTicket);
         // Send response as json
-        response.setContentType(CONTENT_TYPE_JSON);
+        response.setContentType(CesOAuthConstants.CONTENT_TYPE_JSON);
 
         LOGGER.debug("{} : {}", "respond", json.toJSONString());
-        return OAuthUtils.writeText(response, json.toJSONString(), 200);
+        return CesOAuthUtils.writeText(response, json.toJSONString(), 200);
     }
 
     /**
@@ -114,9 +127,9 @@ public final class CesOAuthAccessTokenController extends AbstractController {
         final int expires = (int) (timeout - (System.currentTimeMillis()
                 - ticketGrantingTicket.getCreationTime()) / 1000);
 
-        json.put(OAuthConstants.ACCESS_TOKEN, ticketGrantingTicket.getId());
-        json.put(TOKEN_TYPE, TOKEN_TYPE_VALUE);
-        json.put(TOKEN_EXPIRES, expires);
+        json.put(CesOAuthConstants.ACCESS_TOKEN, ticketGrantingTicket.getId());
+        json.put(CesOAuthConstants.TOKEN_TYPE, CesOAuthConstants.TOKEN_TYPE_VALUE);
+        json.put(CesOAuthConstants.TOKEN_EXPIRES, expires);
 
         return json;
     }
@@ -133,7 +146,7 @@ public final class CesOAuthAccessTokenController extends AbstractController {
     private boolean verifyAccessTokenRequest(final String redirectUri,
                                              final String clientId, final String clientSecret, final String code) {
 
-        final OAuthRegisteredService service = OAuthUtils.getRegisteredOAuthService(this.servicesManager, clientId);
+        final CesOAuthRegisteredService service = CesOAuthUtils.getRegisteredOAuthService(this.servicesManager, clientId);
 
         return verifyClientId(clientId) &&
                 verifyRedirctUir(redirectUri) &&
@@ -146,7 +159,7 @@ public final class CesOAuthAccessTokenController extends AbstractController {
 
     private boolean verifyClientId(String clientId) {
         if (StringUtils.isBlank(clientId)) {
-            LOGGER.error("Missing {}", OAuthConstants.CLIENT_ID);
+            LOGGER.error("Missing {}", CesOAuthConstants.CLIENT_ID);
             return false;
         }
         return true;
@@ -154,7 +167,7 @@ public final class CesOAuthAccessTokenController extends AbstractController {
 
     private boolean verifyRedirctUir(String redirectUri) {
         if (StringUtils.isBlank(redirectUri)) {
-            LOGGER.error("Missing {}", OAuthConstants.REDIRECT_URI);
+            LOGGER.error("Missing {}", CesOAuthConstants.REDIRECT_URI);
             return false;
         }
         return true;
@@ -162,7 +175,7 @@ public final class CesOAuthAccessTokenController extends AbstractController {
 
     private boolean verifyClientSecret(String clientSecret) {
         if (StringUtils.isBlank(clientSecret)) {
-            LOGGER.error("Missing {}", OAuthConstants.CLIENT_SECRET);
+            LOGGER.error("Missing {}", CesOAuthConstants.CLIENT_SECRET);
             return false;
         }
         return true;
@@ -170,29 +183,29 @@ public final class CesOAuthAccessTokenController extends AbstractController {
 
     private boolean verifyCode(String code) {
         if (StringUtils.isBlank(code)) {
-            LOGGER.error("Missing {}", OAuthConstants.CODE);
+            LOGGER.error("Missing {}", CesOAuthConstants.CODE);
             return false;
         }
         return true;
     }
 
-    private boolean verifyHasRegisteredOAuthService(OAuthRegisteredService service, String clientId) {
+    private boolean verifyHasRegisteredOAuthService(CesOAuthRegisteredService service, String clientId) {
         if (service == null) {
-            LOGGER.error("Unknown {} : {}", OAuthConstants.CLIENT_ID, clientId);
+            LOGGER.error("Unknown {} : {}", CesOAuthConstants.CLIENT_ID, clientId);
             return false;
         }
         return true;
     }
 
-    private boolean verifyRedirectUirMatchesServiceId(OAuthRegisteredService service, String redirectUri) {
+    private boolean verifyRedirectUirMatchesServiceId(CesOAuthRegisteredService service, String redirectUri) {
         if (!redirectUri.matches(service.getServiceId())) {
-            LOGGER.error("Unsupported {} : {} for serviceId : {}", OAuthConstants.REDIRECT_URI, redirectUri, service.getServiceId());
+            LOGGER.error("Unsupported {} : {} for serviceId : {}", CesOAuthConstants.REDIRECT_URI, redirectUri, service.getServiceId());
             return false;
         }
         return true;
     }
 
-    private boolean verifySecretHash(OAuthRegisteredService service, String clientSecret) {
+    private boolean verifySecretHash(CesOAuthRegisteredService service, String clientSecret) {
         // clientSecretHash is a hash from cas over the clientSecretHash, therefore we need to encrypt the clientSecretHash to perform the check
         String clientSecretHash = org.apache.commons.codec.digest.DigestUtils.sha256Hex(clientSecret);
         if (!StringUtils.equals(service.getClientSecret(), clientSecretHash)) {