+++ app/src/test/java/de/triology/cas/services/CesServicesManagerStageProductiveTest.java
@@ -1,7 +1,8 @@
 package de.triology.cas.services;
 
+import de.triology.cas.oauth.service.CesOAuthServiceFactory;
 import de.triology.cas.services.Registry.DoguChangeListener;
-import de.triology.cas.services.oauth.CesOAuthServiceFactory;
+import de.triology.cas.services.dogu.CesDoguServiceFactory;
 import org.jasig.cas.services.RegisteredService;
 import org.junit.Assert;
 import org.junit.Before;